♻️:: activity 기능 UseCase 패키징 변경 및 Entity, Param 변경사항 적용
