+++ core/domain/src/main/java/com/msg/domain/usecase/activity/ApproveStudentActivityInfoUseCase.kt
@@ -1,13 +1,13 @@
-package com.msg.domain.activity
+package com.msg.domain.usecase.activity
 
 import com.msg.data.repository.activity.ActivityRepository
 import java.util.UUID
 import javax.inject.Inject
 
-class DeleteStudentActivityInfoUseCase @Inject constructor(
+class ApproveStudentActivityInfoUseCase @Inject constructor(
     private val activityRepository: ActivityRepository
 ) {
     suspend operator fun invoke(id: UUID) = runCatching {
-        activityRepository.deleteStudentActivityInfo(id = id)
+        activityRepository.approveStudentActivityInfo(id = id)
     }
 }
