+++ core/domain/src/main/java/com/msg/domain/usecase/activity/ChangePasswordUseCase.kt
@@ -1,13 +1,13 @@
-package com.msg.domain.user
+package com.msg.domain.usecase.activity
 
 import com.msg.data.repository.user.UserRepository
-import com.msg.model.remote.request.user.ChangePasswordRequest
+import com.msg.model.param.user.ChangePasswordParam
 import javax.inject.Inject
 
 class ChangePasswordUseCase @Inject constructor(
     private val userRepository: UserRepository
 ) {
-    suspend operator fun invoke(body: ChangePasswordRequest) = runCatching {
+    suspend operator fun invoke(body: ChangePasswordParam) = runCatching {
         userRepository.changePassword(body = body)
     }
 }
