+++ core/domain/src/main/java/com/msg/domain/usecase/activity/GetDetailStudentActivityInfoUseCase.kt
@@ -1,13 +1,13 @@
-package com.msg.domain.activity
+package com.msg.domain.usecase.activity
 
 import com.msg.data.repository.activity.ActivityRepository
 import java.util.UUID
 import javax.inject.Inject
 
-class RejectStudentActivityInfoUseCase @Inject constructor(
+class GetDetailStudentActivityInfoUseCase @Inject constructor(
     private val activityRepository: ActivityRepository
 ) {
     suspend operator fun invoke(id: UUID) = runCatching {
-        activityRepository.rejectStudentActivityInfo(id = id)
+        activityRepository.getDetailStudentActivityInfo(id = id)
     }
 }
