+++ core/domain/src/main/java/com/msg/domain/usecase/activity/GetMyStudentActivityInfoListUseCase.kt
@@ -1,4 +1,4 @@
-package com.msg.domain.activity
+package com.msg.domain.usecase.activity
 
 import com.msg.data.repository.activity.ActivityRepository
 import javax.inject.Inject
