+++ core/domain/src/main/java/com/msg/domain/usecase/activity/RejectStudentActivityInfoUseCase.kt
@@ -1,13 +1,13 @@
-package com.msg.domain.activity
+package com.msg.domain.usecase.activity
 
 import com.msg.data.repository.activity.ActivityRepository
 import java.util.UUID
 import javax.inject.Inject
 
-class ApproveStudentActivityInfoUseCase @Inject constructor(
+class RejectStudentActivityInfoUseCase @Inject constructor(
     private val activityRepository: ActivityRepository
 ) {
     suspend operator fun invoke(id: UUID) = runCatching {
-        activityRepository.approveStudentActivityInfo(id = id)
+        activityRepository.rejectStudentActivityInfo(id = id)
     }
 }
